Move deck functionality into its own module. Deck is a struct with associated functions.
